Offloaded some function from script js
