+++ blocks/card-list/libraryCard.js
@@ -1,5 +1,5 @@
-import { makePublicUrl } from '../../scripts/scripts.js';
-import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
+import { makePublicUrl } from '../../scripts/url-helper.js';
+import { createOptimizedPicture } from '../../scripts/image-helper.js';
 import {
   li, a, p, div, h2,
 } from '../../scripts/dom-builder.js';
