+++ blocks/cards/cards.js
@@ -1,8 +1,8 @@
-import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
+import { createOptimizedPicture } from '../../scripts/image-helper.js';
 import {
   a, div, li, ul,
 } from '../../scripts/dom-builder.js';
-import { makePublicUrl } from '../../scripts/scripts.js';
+import { makePublicUrl } from '../../scripts/url-helper.js';
 
 export default function decorate(block) {
   /* change to ul, li */
