+++ blocks/product-category-list/product-category-list.js
@@ -2,7 +2,8 @@ import ffetch from '../../scripts/ffetch.js';
 import {
   ul, li, a, h3, span,
 } from '../../scripts/dom-builder.js';
-import { makePublicUrl, imageHelper } from '../../scripts/scripts.js';
+import { makePublicUrl } from '../../scripts/url-helper.js';
+import { imageHelper } from '../../scripts/image-helper.js';
 import { toClassName, getMetadata } from '../../scripts/lib-franklin.js';
 import createFilters from './filter.js';
