+++ blocks/product-category/product-category.js
@@ -2,7 +2,8 @@ import ffetch from '../../scripts/ffetch.js';
 import {
   ul, li, a, p, div, span, h4,
 } from '../../scripts/dom-builder.js';
-import { makePublicUrl, imageHelper } from '../../scripts/scripts.js';
+import { makePublicUrl } from '../../scripts/url-helper.js';
+import { imageHelper } from '../../scripts/image-helper.js';
 import { getMetadata } from '../../scripts/lib-franklin.js';
 import { buildProductCategorySchema } from '../../scripts/schema.js';
