+++ blocks/recent-articles/recent-articles.js
@@ -2,7 +2,8 @@ import ffetch from '../../scripts/ffetch.js';
 import {
   div, ul, li, a, p, span,
 } from '../../scripts/dom-builder.js';
-import { formatDateUTCSeconds, makePublicUrl } from '../../scripts/scripts.js';
+import { makePublicUrl } from '../../scripts/url-helper.js';
+import { formatDateUTCSeconds } from '../../scripts/date-helper.js';
 import { getMetadata } from '../../scripts/lib-franklin.js';
 
 export default async function decorate(block) {
